+++ cmd/ttn-lw-cli/commands/flags.go
@@ -23,6 +23,7 @@ import (
 	"github.com/spf13/pflag"
 	"go.thethings.network/lorawan-stack/pkg/errors"
 	"go.thethings.network/lorawan-stack/pkg/ttnpb"
+	"google.golang.org/grpc"
 )
 
 func firstArgs(i int, args ...string) []string {
@@ -134,20 +135,52 @@ func searchFlags() *pflag.FlagSet {
 	flagSet.String("name-contains", "", "")
 	flagSet.String("description-contains", "", "")
 	flagSet.StringToString("attributes-contain", nil, "(key=value)")
+	flagSet.AddFlagSet(paginationFlags())
 	return flagSet
 }
 
-func getSearchEntitiesRequest(flagSet *pflag.FlagSet) *ttnpb.SearchEntitiesRequest {
+func getSearchEntitiesRequest(flagSet *pflag.FlagSet) (req *ttnpb.SearchEntitiesRequest, opt grpc.CallOption, getTotal func() uint64) {
 	idContains, _ := flagSet.GetString("id-contains")
 	nameContains, _ := flagSet.GetString("name-contains")
 	descriptionContains, _ := flagSet.GetString("description-contains")
 	attributesContain, _ := flagSet.GetStringToString("attributes-contain")
+	limit, page, opt, getTotal := withPagination(flagSet)
 	return &ttnpb.SearchEntitiesRequest{
 		IDContains:          idContains,
 		NameContains:        nameContains,
 		DescriptionContains: descriptionContains,
 		AttributesContain:   attributesContain,
-	}
+		Limit:               limit,
+		Page:                page,
+	}, opt, getTotal
+}
+
+func searchEndDevicesFlags() *pflag.FlagSet {
+	flagSet := &pflag.FlagSet{}
+	flagSet.String("dev-eui-contains", "", "")
+	flagSet.String("join-eui-contains", "", "")
+	flagSet.String("dev-addr-contains", "", "")
+	flagSet.Lookup("dev-addr-contains").Hidden = true // Part of the API but not actually supported.
+	flagSet.AddFlagSet(searchFlags())
+	return flagSet
+}
+
+func getSearchEndDevicesRequest(flagSet *pflag.FlagSet) (req *ttnpb.SearchEndDevicesRequest, opt grpc.CallOption, getTotal func() uint64) {
+	baseReq, opt, getTotal := getSearchEntitiesRequest(flagSet)
+	devEUIContains, _ := flagSet.GetString("dev-eui-contains")
+	joinEUIContains, _ := flagSet.GetString("join-eui-contains")
+	devAddrContains, _ := flagSet.GetString("dev-addr-contains")
+	return &ttnpb.SearchEndDevicesRequest{
+		IDContains:          baseReq.IDContains,
+		NameContains:        baseReq.NameContains,
+		DescriptionContains: baseReq.DescriptionContains,
+		AttributesContain:   baseReq.AttributesContain,
+		DevEUIContains:      devEUIContains,
+		JoinEUIContains:     joinEUIContains,
+		DevAddrContains:     devAddrContains,
+		Limit:               baseReq.Limit,
+		Page:                baseReq.Page,
+	}, opt, getTotal
 }
 
 var errNoIDs = errors.DefineInvalidArgument("no_ids", "no IDs set")