+++ cmd/ttn-lw-cli/commands/organizations.go
@@ -94,17 +94,18 @@ var (
 		RunE: func(cmd *cobra.Command, args []string) error {
 			paths := util.SelectFieldMask(cmd.Flags(), selectOrganizationFlags)
 
-			req := getSearchEntitiesRequest(cmd.Flags())
+			req, opt, getTotal := getSearchEntitiesRequest(cmd.Flags())
 			req.FieldMask.Paths = paths
 
 			is, err := api.Dial(ctx, config.IdentityServerGRPCAddress)
 			if err != nil {
 				return err
 			}
-			res, err := ttnpb.NewEntityRegistrySearchClient(is).SearchOrganizations(ctx, req)
+			res, err := ttnpb.NewEntityRegistrySearchClient(is).SearchOrganizations(ctx, req, opt)
 			if err != nil {
 				return err
 			}
+			getTotal()
 
 			return io.Write(os.Stdout, config.OutputFormat, res.Organizations)
 		},