+++ pkg/identityserver/store/store_interfaces.go
@@ -169,7 +169,8 @@ type InvitationStore interface {
 
 // EntitySearch interface for searching entities.
 type EntitySearch interface {
-	FindEntities(ctx context.Context, req *ttnpb.SearchEntitiesRequest, entityType string) ([]ttnpb.Identifiers, error)
+	FindEntities(ctx context.Context, member *ttnpb.OrganizationOrUserIdentifiers, req *ttnpb.SearchEntitiesRequest, entityType string) ([]ttnpb.Identifiers, error)
+	FindEndDevices(ctx context.Context, req *ttnpb.SearchEndDevicesRequest) ([]*ttnpb.EndDeviceIdentifiers, error)
 }
 
 // ContactInfoStore interface for contact info validation.