+++ CHANGELOG.md
@@ -13,6 +13,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add support for Redis Sentinel (see `redis.failover.enable`, `redis.failover.master-name`, `redis.failover.addresses` options).
 - Add options `is.oauth.ui.branding-base-url` and `console.ui.branding-base-url` that can be used to customize the branding (logos) of the web UI.
 - Email templates can now also be loaded from blob buckets.
+- Support for pagination in search APIs.
+- Search is now also available to non-admin users.
+- Support for searching end devices within an application.
 
 ### Changed